feat : ground 리팩토링 중 create, delete 로직 보강 및 테스트

+++ src/main/java/companion/challeculum/domains/ground/GroundServiceImpl.java
@@ -1,12 +1,15 @@
 package companion.challeculum.domains.ground;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import companion.challeculum.common.Constants;
+import companion.challeculum.common.UpdateRecordUtil;
 import companion.challeculum.domains.ground.dtos.Ground;
 import companion.challeculum.domains.ground.dtos.GroundCreateDto;
 import companion.challeculum.domains.ground.dtos.GroundJoined;
 import companion.challeculum.domains.ground.dtos.GroundUpdateDto;
 import companion.challeculum.domains.mission.MissionDao;
+import companion.challeculum.domains.user.UserDao;
+import companion.challeculum.domains.user.dtos.User;
+import companion.challeculum.domains.user.dtos.UserUpdateDto;
 import companion.challeculum.domains.userground.UserGroundDao;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,6 +19,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Function;
 
 import static companion.challeculum.common.Constants.ROWS_PER_PAGE;
 
@@ -26,63 +30,80 @@ public class GroundServiceImpl implements GroundService {
     private final GroundDao groundDao;
     private final MissionDao missionDao;
     private final UserGroundDao userGroundDao;
+    private final UpdateRecordUtil updateRecordUtil;
+    private final UserDao userDao;
 
     @Override
-    public void deleteGround(long groundId) {
+    public void deleteGround(User user, long groundId) {
         Ground ground = groundDao.getGroundByGroundId(groundId);
         if (!ground.getStatus().equals(Constants.GROUND_STANDBY)) throw new RuntimeException(); // 에러 핸들링
+        if (user.getUserId() != ground.getCreateUserId() && user.getRole().equals(Constants.ROLE_MEMBER)) {
+            throw new RuntimeException(); // 에러 핸들링
+        }
+        userGroundDao.getUserGroundListByGroundId(groundId).forEach(userGround -> {
+            UserUpdateDto userUpdateDto = new UserUpdateDto();
+            userUpdateDto.setPoint(userUpdateDto.getPoint() + ground.getDeposit());
+            Map<String, Object> newUpdatedMap = updateRecordUtil.generateUpdateMap(userUpdateDto,
+                    Constants.TO_SNAKE_CASE, Function.identity());
+            userDao.update(userGround.getUserId(), newUpdatedMap);
+        });
         groundDao.deleteGround(groundId);
     }
 
 
     @Override
     public List<GroundJoined> getGroundList(Integer page, String filter, String sortBy, String orderBy, String keyword) {
-        Integer startRow = ROWS_PER_PAGE * (page - 1);
+        int startRow = ROWS_PER_PAGE * (page - 1);
         Map<String, String> filterMap = new HashMap<>();
         String[] pairs = filter.split(",");
         Arrays.stream(pairs).filter(p -> !(p.equals(""))).forEach(p ->{
             String[] keyValue = p.split(":");
             if (!keyValue[1].equals("0") || !keyValue[1].equals("undefiend")) filterMap.put(keyValue[0], keyValue[1]);
         });
-        return groundDao.getGroundList(startRow, ROWS_PER_PAGE, filterMap, sortBy, orderBy, keyword);
+        List<GroundJoined> groundJoinedList = groundDao.getGroundList(startRow, ROWS_PER_PAGE, filterMap, sortBy, orderBy, keyword);
+        return fecthJoinMissionList(groundJoinedList);
 
     }
 
+
     @Override
     public List<GroundJoined> getGroundsByMe(long userId) {
-        return groundDao.getGroundsByMe(userId);
+        List<GroundJoined> groundJoinedList = groundDao.getGroundsByMe(userId);
+        return fecthJoinMissionList(groundJoinedList);
     }
 
     @Override
     public void updateGround(long groundId, GroundUpdateDto groundUpdateDto) {
-        ObjectMapper objectMapper = new ObjectMapper();
-        Map<String, Object> updateMap = objectMapper.convertValue(groundUpdateDto, Map.class);
-        Map<String, Object> newUpdatedMap = new HashMap<>();
-
-        for (Map.Entry<String, Object> entry : updateMap.entrySet()) {
-            String key = entry.getKey();
-            String regex = "([a-z])([A-Z])";
-            String replacement = "$1_$2";
-            String newKey = key.replaceAll(regex, replacement).toLowerCase();
-            newUpdatedMap.put(newKey, entry.getValue());
-        }
+        Map<String, Object> newUpdatedMap = updateRecordUtil.generateUpdateMap(groundUpdateDto,
+                Constants.TO_SNAKE_CASE, Function.identity());
         groundDao.update(groundId, newUpdatedMap);
     }
 
     @Override
     public List<GroundJoined> getMyGrounds(long userId) {
-        return groundDao.getMyGrounds(userId);
+        List<GroundJoined> groundJoinedList = groundDao.getMyGrounds(userId);
+        return fecthJoinMissionList(groundJoinedList);
     }
 
     //    Redesign(2/25)
     @Override
-    public void createGround(GroundCreateDto dto) {
+    public void createGround(User user, GroundCreateDto dto) {
+        long userId = user.getUserId();
+        int userPoint = user.getPoint();
+        if (dto.getDeposit() > userPoint) throw new RuntimeException(); // 애러 핸들링
+        dto.setCreateUserId(userId);
         long groundId = groundDao.insert(dto);
         dto.getMissionList().forEach(missionCreateDto -> {
             missionCreateDto.setGroundId(groundId);
             missionDao.insert(missionCreateDto);
         });
         userGroundDao.insert(dto.getCreateUserId(), groundId);
+        UserUpdateDto userUpdateDto = new UserUpdateDto();
+        userUpdateDto.setPoint(userPoint - dto.getDeposit());
+        Map<String, Object> newUpdatedMap = updateRecordUtil.generateUpdateMap(userUpdateDto,
+                Constants.TO_SNAKE_CASE, Function.identity());
+        userDao.update(userId, newUpdatedMap);
+
     }
 
     @Override
@@ -90,9 +111,17 @@ public class GroundServiceImpl implements GroundService {
         return groundDao.getGroundByGroundId(groundId);
     }
 
-
     @Override
     public GroundJoined getGroundJoinedByGroundId(long groundId) {
-        return groundDao.getGroundJoinedByGroundId(groundId);
+        GroundJoined groundJoined = groundDao.getGroundJoinedByGroundId(groundId);
+        int numOfParticipants = userGroundDao.getUserGroundJoinedListByGroundId(groundId).size();
+        groundJoined.setNumOfParticipants(numOfParticipants);
+        groundJoined.setMissionList(missionDao.getMissionList(groundId));
+        return groundJoined;
+    }
+    private List<GroundJoined> fecthJoinMissionList(List<GroundJoined> groundJoinedList) {
+        return groundJoinedList.stream()
+                .peek(groundJoined -> groundJoined.setMissionList(missionDao.getMissionList(groundJoined.getGroundId())))
+                .toList();
     }
 }
